ResetPassword: consolidate ResetPasswordHeader
